src/content/events/requirements.txt
boto3
python-dotenv
rich
